+++ UltimateRedditBot.Domain/Models/GuildSettings.cs
@@ -5,7 +5,7 @@ namespace UltimateRedditBot.Domain.Models
     public class GuildSettings : BaseEntity
     {
         public Guild Guild { get; set; }
-        public int GuildId { get; set; }
+        public ulong GuildId { get; set; }
 
         public string Key { get; set; }
         public string Value { get; set; }
